Refactor: Modernize packaging and begin codebase refactor
- Updated pyproject.toml for Hatch and UV usage.
- Updated GitHub Actions to use Hatch for build and test.
- Updated dependencies: NumPy, Ruff, Mypy, Pytest.
- Created fast_mypyc and fast_numba subdirectories.
- Moved Mypyc candidates and Numba files to respective subdirectories.
- Updated Mypyc configuration in pyproject.toml.
- Added __init__.py files to new subdirectories.
- Partially updated import paths (e.g., in cli.py).

Further work is needed to fully update all import paths across the codebase and tests.
