+++ src/imgcolorshine/fast_mypyc/__init__.py
@@ -1,0 +1,2 @@
+# this_file: src/imgcolorshine/fast_mypyc/__init__.py
+"""Mypyc-compiled modules for imgcolorshine."""
