+++ src/imgcolorshine/fast_mypyc/colorshine.py
