+++ src/imgcolorshine/fast_mypyc/engine.py
