+++ src/imgcolorshine/fast_mypyc/falloff.py
