+++ src/imgcolorshine/fast_mypyc/io.py
