+++ src/imgcolorshine/fast_numba/__init__.py
@@ -1,0 +1,2 @@
+# this_file: src/imgcolorshine/fast_numba/__init__.py
+"""Numba-accelerated modules for imgcolorshine."""
