+++ src/imgcolorshine/fast_numba/numba_utils.py
